Make assertions more based on DOM rather than text
A similar idea can be applied to all of the tests. I'm slightly torn
here - what is already here works for how our trainees are likely to
solve these problems, but the tests are quite brittle and I probably
wouldn't be happy with them IRL (because they wouldn't account for e.g.
wrapping divs around elements, extra properties in-line as the code
changes, newlines in the generated code, etc).